requirements.txt
streamlit
python-dotenv
requests
pyyaml
feedparser
